+++ components/misc/NotAvailable.tsx
@@ -2,18 +2,19 @@ import React from "react";
 import Typography from '@material-ui/core/Typography';
 import MiddleEllipsis from './MiddleEllipsis'
 
+type NotAvailableProps = { wrap?: boolean, variant: string, className?: string };
 
-const NotAvailable = (props: any) => {
-    if (props.wrap) {
+const NotAvailable = ({ wrap, variant, className }: NotAvailableProps) => {
+    if (wrap) {
         return (
-            <Typography variant={props.variant} className={props.className}  >
+            <Typography variant={variant} className={className}  >
                 <MiddleEllipsis text="Data not available" />
             </Typography>
         )
 
     }
     return (
-        <Typography variant={props.variant} className={props.className}  >
+        <Typography variant={variant} className={className} noWrap>
             Data not available
         </Typography>
     )